requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
kafka-python==1.4.6
psycopg2-binary==2.8.3
requests==2.22.0
urllib3==1.25.3
